% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bankroll_plot.R
\name{bankroll_plot}
\alias{bankroll_plot}
\title{Bankroll Plot}
\usage{
bankroll_plot(prev_months = 24, league = "ALL")
}
\arguments{
\item{prev_months}{Numeric - Number of Months back from today that will be included}

\item{league}{String - Leagues to include in the analysis}
}
\value{
A tibble - The most profitable days betting
}
\description{
Bankroll Plot
}
\examples{
bankroll_plot(months = 5)
}
