% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_days.R
\name{best_days}
\alias{best_days}
\title{Top betting days}
\usage{
best_days(days = 10)
}
\arguments{
\item{days}{Numeric - Number of days}
}
\value{
A tibble - The most profitable days betting
}
\description{
Top betting days
}
\examples{
best_days(5)
}
