% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worst_days.R
\name{worst_days}
\alias{worst_days}
\title{Worst betting days}
\usage{
worst_days(days = 10)
}
\arguments{
\item{days}{Numeric - Number of days}
}
\value{
A tibble - The least profitable days betting
}
\description{
Worst betting days
}
\examples{
worst_days(5)
}
